Move types from lib to types folder
