+++ app/api/top-tracks/route.ts
@@ -1,5 +1,5 @@
 import { getTopTracks, spotifyResponse } from 'lib/spotify'
-import { ReadableTrack, Track } from 'lib/types'
+import { ReadableTrack, Track } from 'types/spotify'
 
 export const runtime = 'edge';
 export const fetchCache = 'force-no-store';
